Optimize CRC32 allocations for IOs
Reuse the string buffer, and if the
the Ruby version supports it preallocate
the correct string capacity upfront
