+++ lib/zip_tricks/stream_crc32.rb
@@ -2,13 +2,26 @@
 
 # A simple stateful class for keeping track of a CRC32 value through multiple writes
 class ZipTricks::StreamCRC32
+  STRINGS_HAVE_CAPACITY_SUPPORT = begin
+    String.new('', capacity: 1)
+    true
+  rescue ArgumentError
+    false
+  end
+  CRC_BUF_SIZE = 1024 * 512
+  private_constant :STRINGS_HAVE_CAPACITY_SUPPORT, :CRC_BUF_SIZE
+
   # Compute a CRC32 value from an IO object. The object should respond to `read` and `eof?`
   #
   # @param io[IO] the IO to read the data from
   # @return [Fixnum] the computed CRC32 value
   def self.from_io(io)
+    # If we can specify the string capacity upfront we will not have to resize
+    # the string during operation. This saves time but is only available on
+    # recent Ruby 2.x versions.
+    blob = STRINGS_HAVE_CAPACITY_SUPPORT ? String.new('', capacity: CRC_BUF_SIZE) : String.new('')
     crc = new
-    crc << io.read(1024 * 512) until io.eof?
+    crc << io.read(CRC_BUF_SIZE, blob) until io.eof?
     crc.to_i
   end
